load woff files lazy upon export
